Make dispense transaction more normal
